Continue stripping
But the exception already seems less likely, even though all we did is inline
stuff...
