--- test-grapesy/Test/Driver/Dialogue.hs
@@ -1,10 +1,0 @@
-{-# LANGUAGE OverloadedStrings #-}
-
-module Test.Driver.Dialogue (
-    module X
-  ) where
-
-import Test.Driver.Dialogue.Definition as X
-import Test.Driver.Dialogue.Execution  as X
-import Test.Driver.Dialogue.Generation as X
-import Test.Driver.Dialogue.TestClock  as X
